# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CryoET Catalog API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/tomograms/{sample_id}/{acquisition_id}/{reconstruction_alignment_id}/{tomogram_id}/preview.png": {
      "get": {
        "tags": [
          "tomograms"
        ],
        "summary": "Tomogram Preview",
        "description": "Render the center-XY slice as a PNG (1200px wide, 1\u201399% percentile).\n\nReturns 404 for missing row or path-outside-root, 422 for an existing\nrow whose ``mrc_path`` is missing on disk.",
        "operationId": "tomogram_preview_tomograms__sample_id___acquisition_id___reconstruction_alignment_id___tomogram_id__preview_png_get",
        "parameters": [
          {
            "name": "sample_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sample Id"
            }
          },
          {
            "name": "acquisition_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Acquisition Id"
            }
          },
          {
            "name": "reconstruction_alignment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Reconstruction Alignment Id"
            }
          },
          {
            "name": "tomogram_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tomogram Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tomograms/{sample_id}/{acquisition_id}/{reconstruction_alignment_id}/{tomogram_id}/neuroglancer": {
      "post": {
        "tags": [
          "tomograms"
        ],
        "summary": "Tomogram Neuroglancer",
        "description": "Launch a Neuroglancer viewer over the tomogram volume.\n\nThe frontend rewrites the URL hostname to ``window.location.hostname``\nbefore opening \u2014 Neuroglancer reports the API host's FQDN which may\nnot be reachable from the browser.",
        "operationId": "tomogram_neuroglancer_tomograms__sample_id___acquisition_id___reconstruction_alignment_id___tomogram_id__neuroglancer_post",
        "parameters": [
          {
            "name": "sample_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sample Id"
            }
          },
          {
            "name": "acquisition_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Acquisition Id"
            }
          },
          {
            "name": "reconstruction_alignment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Reconstruction Alignment Id"
            }
          },
          {
            "name": "tomogram_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tomogram Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ViewerLaunchOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "ViewerLaunchOut": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url"
          }
        },
        "type": "object",
        "required": [
          "url"
        ],
        "title": "ViewerLaunchOut",
        "description": "Response of a POST .../neuroglancer launch.\n\nFrontend rewrites the hostname to ``window.location.hostname`` before\nopening."
      },
...
    }
  }
}